refactor: UserDataRepository and UserRepository divide
