+++ src/app/components/Authtorization/credentials_auth_form/CredentialsAuthForm.container.tsx
@@ -7,7 +7,9 @@ import { IUser } from "@/app/modules/Auth/domain/user_credentials/IUser";
 import { useRouter } from "next/navigation";
 
 export default function CredentialsAuthFormContainer() {
-  const repository = diContainer.getUserRepository();
+  const authRepository = diContainer.getAuthRepository();
+  const userDataRepository = diContainer.getUserDataRepository();
+
   const router = useRouter();
 
   const [formData, setFormData] = React.useState({
@@ -26,8 +28,8 @@ export default function CredentialsAuthFormContainer() {
   const handleSubmit = async (e: any) => {
     e.preventDefault();
 
-    repository.getToken(formData.email, formData.password);
-    const userData: IUser = await repository.getUserData();
+    authRepository.getToken(formData.email, formData.password);
+    const userData: IUser = await userDataRepository.getUserData();
     console.log(userData);
 
     router.push("/pages/profilePage");
