+++ src/app/di/AppDiContainer.ts
@@ -1,13 +1,13 @@
-import { UserRepositoryImpl } from "../modules/Auth/data/user_credentials/UserRepositoryImpl";
-import { UserRepository } from "../modules/Auth/domain/user_credentials/UserRepository";
+import { AuthRepositoryImpl } from "../modules/Auth/data/user_credentials/AuthRepositoryImpl";
+import { AuthRepository } from "../modules/Auth/domain/user_credentials/AuthRepository";
 import { UserDataRepositoryImpl } from "../modules/repositories/user_data_repository/data/UserDataRepositoryImpl";
 import { UserDataRepository } from "../modules/repositories/user_data_repository/domain/UserDataRepository";
 
 import { DiContainer } from "./DiContainer";
 
 export class AppDiContainer implements DiContainer {
-  getUserRepository(): UserRepository {
-    return new UserRepositoryImpl();
+  getAuthRepository(): AuthRepository {
+    return new AuthRepositoryImpl();
   }
 
   getUserDataRepository(): UserDataRepository {
