+++ src/app/modules/Auth/domain/user_credentials/AuthRepository.ts
@@ -1,0 +1,3 @@
+export interface AuthRepository {
+  getToken(email: string, password: string): Promise<void>;
+}
