--- src/app/modules/Auth/domain/user_credentials/UserRepository.ts
@@ -1,10 +1,0 @@
-import { IToken } from "./IToken";
-import { IUser } from "./IUser";
-
-export interface UserRepository {
-  getToken(email: string, password: string): Promise<void>;
-
-  getUserData(): Promise<IUser>;
-
-  getTokenFromLocalStorage(): Promise<IToken>;
-}
