--- src/app/modules/repositories/user_data_repository/data/TokenEntity.ts
@@ -1,3 +1,0 @@
-export interface TokenEntity {
-  token: string;
-}
