+++ src/app/modules/repositories/user_data_repository/data/UserEntity.ts
@@ -1,0 +1,6 @@
+export interface UserEntity {
+  id: number;
+  username: string;
+  email: string;
+  photo: string;
+}
