+++ src/app/modules/repositories/user_data_repository/domain/User.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: number;
+  username: string;
+  email: string;
+  photo: string;
+}
